% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im.import.R
\name{im.import}
\alias{im.import}
\title{Import a raster image}
\usage{
im.import(im)
}
\arguments{
\item{im}{A character string. Either the name of an image included in the package, 
or a full file path to a user-provided raster image.}
}
\value{
A `SpatRaster` object.
}
\description{
This function imports a raster image from the package's internal image collection,
from a user-specified local path, or from a remote Zenodo repository if the image
is not found locally.
}
