\name{im.print {imageRy}}
\alias{im.print}
\title{
Prints a character line
}
\description{
The function will print "I am imageRy".
}
\usage{
im.print()
}
\references{
Go check print() function.
}
\author{
Duccio Rocchini
}
\seealso{
https://github.com/ducciorocchini/imageRy/
}
\examples{
im.print()
