\name{im.refresh()}
\alias{im.refresh()}
\title{
Re-install imageRy package 
}
\description{
This function forces a re-installation of the "imageRy" package from the official github repository.
}
\usage{
im.refresh()
}
\author{
Duccio Rocchini
}
\seealso{
https://github.com/ducciorocchini/imageRy/
}
\examples{
im.refresh()
}
