\name{imgTest_2}
\docType{data}
\alias{t2}
\title{Amazon deforestation: t2}
\description{
  This data set gives a RasterBrick object representing Sentinel-2 captures during the last 6 mounths of 2019.
  Blue, Green, Red and NIR bands with pixel resolution equal to 10x10 meters.
}
\usage{t2}
\format{A RasterBrick containing 27482 cells, 4 layers.}
\source{Harmonized Sentinel-2 MSI: MultiSpectral Instrument, Level-2A.}
\keyword{datasets}
