\name{imgTest_3}
\docType{data}
\alias{t3}
\title{Amazon deforestation: t3}
\description{
  This data set gives a RasterBrick object representing Sentinel-2 captures during the first 6 mounths of 2020.
  Blue, Green, Red and NIR bands with pixel resolution equal to 10x10 meters.
}
\usage{t3}
\format{A RasterBrick containing 27482 cells, 4 layers.}
\source{Harmonized Sentinel-2 MSI: MultiSpectral Instrument, Level-2A.}
\keyword{datasets}
