\name{imgTest_4}
\docType{data}
\alias{t4}
\title{Amazon deforestation: t4}
\description{
  This data set gives a RasterBrick object representing Sentinel-2 captures during the last 6 mounths of 2020.
  Blue, Green, Red and NIR bands with pixel resolution equal to 10x10 meters.
}
\usage{t4}
\format{A RasterBrick containing 27482 cells, 4 layers.}
\source{Harmonized Sentinel-2 MSI: MultiSpectral Instrument, Level-2A.}
\keyword{datasets}
