% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im.import.R
\name{im.import}
\alias{im.import}
\title{Import one or more raster images}
\usage{
im.import(im)
}
\arguments{
\item{im}{Character vector. Names of images included in the package,
or full file paths to user-provided raster images.}
}
\value{
A `SpatRaster` object (stack of layers).
}
\description{
This function imports raster images from the package's internal collection,
from user-specified local paths, or from the remote Zenodo repository.
}
