\name{im.list {imageRy}}
\alias{im.list}
\title{
List the Files in imageRy
}
\description{
This function produces a character vector of the names of files stored in the imageRy package directory.
}
\usage{
im.list()
}
\value{
A character vector.
}
\author{
Elisa Thouverai, Duccio Rocchini
}
\seealso{
https://github.com/ducciorocchini/imageRy/
}
\examples{
im.list()
