% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im.list.R
\name{im.list}
\alias{im.list}
\title{List Available Raster Images in the imageRy Package}
\usage{
im.list()
}
\value{
A character vector containing the names of available raster image files.
}
\description{
This function lists all raster images stored in the `imageRy` package.
}
\details{
The function retrieves the names of all files in the "images" directory of the `imageRy` package.
These files can be imported using `im.import()`.
}
\examples{
\dontrun{
library(imageRy)

# List available images
im.list()
}

}
\seealso{
[im.import()]
}
