% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im.print.R
\name{im.print}
\alias{im.print}
\title{Print a Message Identifying the imageRy Package}
\usage{
im.print()
}
\value{
This function does not return an object. It prints a message to the console.
}
\description{
This function prints a simple message to indicate that the `imageRy` package is being used.
}
\details{
This function serves as a basic test function to verify that the `imageRy` package is loaded.
}
\examples{
\dontrun{
# Print the imageRy message
im.print()
}

}
\seealso{
[im.list()], [im.import()]
}
