% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im.ridgeline.R
\name{im.ridgeline}
\alias{im.ridgeline}
\title{Generate Ridgeline Plots from Satellite Raster Data}
\usage{
im.ridgeline(
  im,
  scale,
  palette = c("viridis", "magma", "plasma", "inferno", "cividis", "mako", "rocket",
    "turbo"),
  direction = 1
)
}
\arguments{
\item{im}{A `SpatRaster` object representing the raster data to be visualized.}

\item{scale}{A numeric value that defines the vertical scale of the ridgeline plot.}

\item{palette}{A character string specifying the `viridis` color palette option to use.
Available options: `"viridis"`, `"magma"`, `"plasma"`, `"inferno"`, `"cividis"`, `"mako"`, `"rocket"`, `"turbo"`.}

\item{direction}{A numeric value controlling the direction of the color scale.
Typically `1` for the default direction and `-1` to reverse the palette.}
}
\value{
A `ggplot` object displaying the ridgeline plot.
}
\description{
This function generates ridgeline plots from stacked satellite imagery data.
}
\details{
Ridgeline plots are useful for analyzing temporal variations in raster-based satellite imagery.
This function extracts raster values and visualizes their distribution across layers.
The `direction` parameter allows reversing the color gradient of the selected palette.
}
\examples{
\dontrun{
library(terra)
library(ggridges)
library(ggplot2)

# Import raster data
r <- im.import("greenland")

# Generate ridgeline plot
im.ridgeline(r, 2, "viridis", direction = 1) + theme_bw()
}
}
\references{
See also `im.import()`, `im.ggplot()`.
}
\seealso{
[GitHub Repository](https://github.com/ducciorocchini/imageRy/)
}
